requirements.txt
pandas
numpy
google-auth
python-dotenv
gitpython
json5
tqdm
docker
openai
numpy
bs4
selenium
tldextract
aiodocker
lzstring
playwright
playwright_stealth
aioboto3
unidecode
validators
inscriptis
tiktoken
pytesseract
dataclasses-json
aiofiles
dill
curlify
w3lib
pydantic
chardet
pytrie
pydig
python-whois
aiohttp
httpx
diff-match-patch
langchain_community
langchain_text_splitters
pygtrie
fastapi[standard]
pyjwt
uvicorn
trafilatura
markdownify
asyncwhois
whodap
termcolor
certstream
websocket-client
PyYAML
certifi
tld
python_Levenshtein
confusables
